requirements.txt
gunicorn==21.2.0
Flask==3.1.0
python-dotenv==1.0.1